buncha changes for code clarity
