+++ .vscode/settings.json
@@ -8,6 +8,7 @@
     "editor.defaultFormatter": "biomejs.biome"
   },
   "editor.codeActionsOnSave": {
-    "source.organizeImports.biome": "explicit"
-  }
+    "source.organizeImports.biome": "always"
+  },
+
 }
