+++ package.json
@@ -17,6 +17,8 @@
     "vite": "^6.2.0"
   },
   "dependencies": {
+    "@types/d3-quadtree": "^3.0.6",
+    "d3-quadtree": "^3.0.1",
     "tiny-invariant": "^1.3.3",
     "zod": "^3.24.4"
   }
